Refactor generated tests into a single file
