--- tests/avro_map.rs
@@ -1,185 +1,0 @@
-
-use converter::convert_avro_direct;
-use serde_json::Value;
-
-#[test]
-fn avro_test_map_with_atomics() {
-    let input_data = r#"
-    {
-      "additionalProperties": {
-        "type": "integer"
-      },
-      "type": "object"
-    }
-    "#;
-    let expected_data = r#"
-    {
-      "name": "root",
-      "type": "map",
-      "values": {
-        "type": "int"
-      }
-    }
-    "#;
-    let input: Value = serde_json::from_str(input_data).unwrap();
-    let expected: Value = serde_json::from_str(expected_data).unwrap();
-    assert_eq!(expected, convert_avro_direct(&input, "root".to_string()));
-}
-
-#[test]
-fn avro_test_map_with_complex() {
-    let input_data = r#"
-    {
-      "additionalProperties": {
-        "properties": {
-          "field_1": {
-            "type": "string"
-          },
-          "field_2": {
-            "type": "integer"
-          }
-        },
-        "type": "object"
-      },
-      "type": "object"
-    }
-    "#;
-    let expected_data = r#"
-    {
-      "name": "root",
-      "type": "map",
-      "values": {
-        "fields": [
-          {
-            "name": "field_1",
-            "type": "string"
-          },
-          {
-            "name": "field_2",
-            "type": "int"
-          }
-        ],
-        "name": "TODO: ???",
-        "type": "record"
-      }
-    }
-    "#;
-    let input: Value = serde_json::from_str(input_data).unwrap();
-    let expected: Value = serde_json::from_str(expected_data).unwrap();
-    assert_eq!(expected, convert_avro_direct(&input, "root".to_string()));
-}
-
-#[test]
-fn avro_test_map_with_pattern_properties() {
-    let input_data = r#"
-    {
-      "additionalProperties": false,
-      "patternProperties": {
-        ".+": {
-          "type": "integer"
-        }
-      },
-      "type": "object"
-    }
-    "#;
-    let expected_data = r#"
-    {
-      "name": "root",
-      "type": "map",
-      "values": {
-        "type": "int"
-      }
-    }
-    "#;
-    let input: Value = serde_json::from_str(input_data).unwrap();
-    let expected: Value = serde_json::from_str(expected_data).unwrap();
-    assert_eq!(expected, convert_avro_direct(&input, "root".to_string()));
-}
-
-#[test]
-fn avro_test_map_with_pattern_and_additional_properties() {
-    let input_data = r#"
-    {
-      "additionalProperties": {
-        "type": "integer"
-      },
-      "patternProperties": {
-        ".+": {
-          "type": "integer"
-        }
-      },
-      "type": "object"
-    }
-    "#;
-    let expected_data = r#"
-    {
-      "name": "root",
-      "type": "map",
-      "values": {
-        "type": "int"
-      }
-    }
-    "#;
-    let input: Value = serde_json::from_str(input_data).unwrap();
-    let expected: Value = serde_json::from_str(expected_data).unwrap();
-    assert_eq!(expected, convert_avro_direct(&input, "root".to_string()));
-}
-
-#[test]
-fn avro_test_incompatible_map_with_pattern_properties() {
-    let input_data = r#"
-    {
-      "additionalProperties": false,
-      "patternProperties": {
-        "^I_": {
-          "type": "integer"
-        },
-        "^S_": {
-          "type": "string"
-        }
-      },
-      "type": "object"
-    }
-    "#;
-    let expected_data = r#"
-    {
-      "name": "root",
-      "type": "map",
-      "values": {
-        "type": "string"
-      }
-    }
-    "#;
-    let input: Value = serde_json::from_str(input_data).unwrap();
-    let expected: Value = serde_json::from_str(expected_data).unwrap();
-    assert_eq!(expected, convert_avro_direct(&input, "root".to_string()));
-}
-
-#[test]
-fn avro_test_incompatible_map_with_pattern_and_additional_properties() {
-    let input_data = r#"
-    {
-      "additionalProperties": {
-        "type": "integer"
-      },
-      "patternProperties": {
-        ".+": {
-          "type": "string"
-        }
-      },
-      "type": "object"
-    }
-    "#;
-    let expected_data = r#"
-    {
-      "name": "root",
-      "type": "map",
-      "values": {
-        "type": "string"
-      }
-    }
-    "#;
-    let input: Value = serde_json::from_str(input_data).unwrap();
-    let expected: Value = serde_json::from_str(expected_data).unwrap();
-    assert_eq!(expected, convert_avro_direct(&input, "root".to_string()));
-}
