--- tests/avro_oneof.rs
@@ -1,424 +1,0 @@
-
-use converter::convert_avro_direct;
-use serde_json::Value;
-
-#[test]
-fn avro_test_oneof_atomic() {
-    let input_data = r#"
-    {
-      "oneOf": [
-        {
-          "type": "integer"
-        },
-        {
-          "type": "integer"
-        }
-      ]
-    }
-    "#;
-    let expected_data = r#"
-    {
-      "name": "root",
-      "type": "int"
-    }
-    "#;
-    let input: Value = serde_json::from_str(input_data).unwrap();
-    let expected: Value = serde_json::from_str(expected_data).unwrap();
-    assert_eq!(expected, convert_avro_direct(&input, "root".to_string()));
-}
-
-#[test]
-fn avro_test_oneof_atomic_with_null() {
-    let input_data = r#"
-    {
-      "oneOf": [
-        {
-          "type": "integer"
-        },
-        {
-          "type": "null"
-        }
-      ]
-    }
-    "#;
-    let expected_data = r#"
-    {
-      "name": "root",
-      "type": [
-        "null",
-        "int"
-      ]
-    }
-    "#;
-    let input: Value = serde_json::from_str(input_data).unwrap();
-    let expected: Value = serde_json::from_str(expected_data).unwrap();
-    assert_eq!(expected, convert_avro_direct(&input, "root".to_string()));
-}
-
-#[test]
-fn avro_test_incompatible_oneof_atomic() {
-    let input_data = r#"
-    {
-      "oneOf": [
-        {
-          "type": "integer"
-        },
-        {
-          "type": "boolean"
-        }
-      ]
-    }
-    "#;
-    let expected_data = r#"
-    {
-      "name": "root",
-      "type": "string"
-    }
-    "#;
-    let input: Value = serde_json::from_str(input_data).unwrap();
-    let expected: Value = serde_json::from_str(expected_data).unwrap();
-    assert_eq!(expected, convert_avro_direct(&input, "root".to_string()));
-}
-
-#[test]
-fn avro_test_incompatible_oneof_atomic_with_null() {
-    let input_data = r#"
-    {
-      "oneOf": [
-        {
-          "type": [
-            "integer",
-            "null"
-          ]
-        },
-        {
-          "type": "boolean"
-        }
-      ]
-    }
-    "#;
-    let expected_data = r#"
-    {
-      "name": "root",
-      "type": [
-        "null",
-        "string"
-      ]
-    }
-    "#;
-    let input: Value = serde_json::from_str(input_data).unwrap();
-    let expected: Value = serde_json::from_str(expected_data).unwrap();
-    assert_eq!(expected, convert_avro_direct(&input, "root".to_string()));
-}
-
-#[test]
-fn avro_test_oneof_object_with_atomics() {
-    let input_data = r#"
-    {
-      "oneOf": [
-        {
-          "properties": {
-            "field_1": {
-              "type": "integer"
-            },
-            "field_2": {
-              "type": "integer"
-            }
-          },
-          "type": "object"
-        },
-        {
-          "properties": {
-            "field_1": {
-              "type": "integer"
-            },
-            "field_2": {
-              "type": "integer"
-            }
-          },
-          "type": "object"
-        }
-      ]
-    }
-    "#;
-    let expected_data = r#"
-    {
-      "fields": [
-        {
-          "name": "field_1",
-          "type": "int"
-        },
-        {
-          "name": "field_2",
-          "type": "int"
-        }
-      ],
-      "name": "root",
-      "type": "record"
-    }
-    "#;
-    let input: Value = serde_json::from_str(input_data).unwrap();
-    let expected: Value = serde_json::from_str(expected_data).unwrap();
-    assert_eq!(expected, convert_avro_direct(&input, "root".to_string()));
-}
-
-#[test]
-fn avro_test_oneof_object_merge() {
-    let input_data = r#"
-    {
-      "oneOf": [
-        {
-          "properties": {
-            "field_1": {
-              "type": "integer"
-            },
-            "field_3": {
-              "type": "number"
-            }
-          },
-          "type": "object"
-        },
-        {
-          "properties": {
-            "field_2": {
-              "type": "boolean"
-            },
-            "field_3": {
-              "type": "number"
-            }
-          },
-          "type": "object"
-        }
-      ]
-    }
-    "#;
-    let expected_data = r#"
-    {
-      "fields": [
-        {
-          "name": "field_1",
-          "type": "int"
-        },
-        {
-          "name": "field_2",
-          "type": "boolean"
-        },
-        {
-          "name": "field_3",
-          "type": "float"
-        }
-      ],
-      "name": "root",
-      "type": "record"
-    }
-    "#;
-    let input: Value = serde_json::from_str(input_data).unwrap();
-    let expected: Value = serde_json::from_str(expected_data).unwrap();
-    assert_eq!(expected, convert_avro_direct(&input, "root".to_string()));
-}
-
-#[test]
-fn avro_test_oneof_object_merge_with_complex() {
-    let input_data = r#"
-    {
-      "oneOf": [
-        {
-          "properties": {
-            "namespace_1": {
-              "properties": {
-                "field_1": {
-                  "type": "integer"
-                },
-                "field_3": {
-                  "type": "number"
-                }
-              },
-              "type": "object"
-            }
-          },
-          "type": "object"
-        },
-        {
-          "properties": {
-            "namespace_1": {
-              "properties": {
-                "field_2": {
-                  "type": "boolean"
-                },
-                "field_3": {
-                  "type": "number"
-                }
-              },
-              "type": "object"
-            }
-          },
-          "type": "object"
-        },
-        {
-          "properties": {
-            "field_4": {
-              "type": "boolean"
-            },
-            "field_5": {
-              "type": "number"
-            }
-          },
-          "type": "object"
-        }
-      ]
-    }
-    "#;
-    let expected_data = r#"
-    {
-      "fields": [
-        {
-          "name": "field_4",
-          "type": "int"
-        },
-        {
-          "name": "field_5",
-          "type": "boolean"
-        },
-        {
-          "fields": [
-            {
-              "name": "field_1",
-              "type": "int"
-            },
-            {
-              "name": "field_2",
-              "type": "boolean"
-            },
-            {
-              "name": "field_3",
-              "type": "float"
-            }
-          ],
-          "name": "namespace_1",
-          "type": "record"
-        }
-      ],
-      "name": "root",
-      "type": "record"
-    }
-    "#;
-    let input: Value = serde_json::from_str(input_data).unwrap();
-    let expected: Value = serde_json::from_str(expected_data).unwrap();
-    assert_eq!(expected, convert_avro_direct(&input, "root".to_string()));
-}
-
-#[test]
-fn avro_test_incompatible_oneof_atomic_and_object() {
-    let input_data = r#"
-    {
-      "oneOf": [
-        {
-          "type": "integer"
-        },
-        {
-          "properties": {
-            "field_1": {
-              "type": "integer"
-            }
-          },
-          "type": "object"
-        }
-      ]
-    }
-    "#;
-    let expected_data = r#"
-    {
-      "name": "root",
-      "type": "string"
-    }
-    "#;
-    let input: Value = serde_json::from_str(input_data).unwrap();
-    let expected: Value = serde_json::from_str(expected_data).unwrap();
-    assert_eq!(expected, convert_avro_direct(&input, "root".to_string()));
-}
-
-#[test]
-fn avro_test_incompatible_oneof_object() {
-    let input_data = r#"
-    {
-      "oneOf": [
-        {
-          "properties": {
-            "field_1": {
-              "type": "integer"
-            }
-          },
-          "type": "object"
-        },
-        {
-          "properties": {
-            "field_1": {
-              "type": "boolean"
-            }
-          },
-          "type": "object"
-        }
-      ]
-    }
-    "#;
-    let expected_data = r#"
-    {
-      "name": "root",
-      "type": "string"
-    }
-    "#;
-    let input: Value = serde_json::from_str(input_data).unwrap();
-    let expected: Value = serde_json::from_str(expected_data).unwrap();
-    assert_eq!(expected, convert_avro_direct(&input, "root".to_string()));
-}
-
-#[test]
-fn avro_test_incompatible_oneof_object_with_complex() {
-    let input_data = r#"
-    {
-      "oneOf": [
-        {
-          "properties": {
-            "namespace_1": {
-              "properties": {
-                "field_1": {
-                  "type": "string"
-                },
-                "field_2": {
-                  "type": "integer"
-                }
-              },
-              "type": "object"
-            }
-          },
-          "type": "object"
-        },
-        {
-          "properties": {
-            "namespace_1": {
-              "properties": {
-                "field_1": {
-                  "type": "boolean"
-                },
-                "field_2": {
-                  "type": "integer"
-                }
-              },
-              "type": "object"
-            }
-          },
-          "type": "object"
-        }
-      ]
-    }
-    "#;
-    let expected_data = r#"
-    {
-      "name": "root",
-      "type": "string"
-    }
-    "#;
-    let input: Value = serde_json::from_str(input_data).unwrap();
-    let expected: Value = serde_json::from_str(expected_data).unwrap();
-    assert_eq!(expected, convert_avro_direct(&input, "root".to_string()));
-}
