--- tests/bigquery_array.rs
@@ -1,65 +1,0 @@
-
-use converter::convert_bigquery_direct;
-use serde_json::Value;
-
-#[test]
-fn bigquery_test_array_with_atomics() {
-    let input_data = r#"
-    {
-      "items": {
-        "type": "integer"
-      },
-      "type": "array"
-    }
-    "#;
-    let expected_data = r#"
-    {
-      "mode": "REPEATED",
-      "type": "INTEGER"
-    }
-    "#;
-    let input: Value = serde_json::from_str(input_data).unwrap();
-    let expected: Value = serde_json::from_str(expected_data).unwrap();
-    assert_eq!(expected, convert_bigquery_direct(&input));
-}
-
-#[test]
-fn bigquery_test_array_with_complex() {
-    let input_data = r#"
-    {
-      "items": {
-        "properties": {
-          "field_1": {
-            "type": "string"
-          },
-          "field_2": {
-            "type": "integer"
-          }
-        },
-        "type": "object"
-      },
-      "type": "array"
-    }
-    "#;
-    let expected_data = r#"
-    {
-      "fields": [
-        {
-          "mode": "NULLABLE",
-          "name": "field_1",
-          "type": "STRING"
-        },
-        {
-          "mode": "NULLABLE",
-          "name": "field_2",
-          "type": "INTEGER"
-        }
-      ],
-      "mode": "REPEATED",
-      "type": "RECORD"
-    }
-    "#;
-    let input: Value = serde_json::from_str(input_data).unwrap();
-    let expected: Value = serde_json::from_str(expected_data).unwrap();
-    assert_eq!(expected, convert_bigquery_direct(&input));
-}
